Backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langchain-community
langgraph
motor
pymongo
dnspython
pydantic
PyJWT
passlib
bcrypt
requests
python-dotenv
python-multipart
email-validator
openai